requirements.txt
sounddevice
numpy
librosa
soundfile
silero-vad
faster-whisper
python-dotenv
